chore: extract the period function here and make it reusable :fire:
